+++ src/DataSet/Material.h
@@ -162,7 +162,7 @@ struct Material_s {           /* material */
   int modelindex ;            /**< Model index */
   
   /* for compatibility with former version (should be eliminated) */
-  unsigned short int neq ;    /**< nombre d'equations du modele */
+  unsigned int neq ;          /**< nombre d'equations du modele */
   char**   eqn ;              /**< nom des equations */
   char**   inc ;              /**< nom des inconnues */
   int      nc ;               /**< nb of curves */
